Parse a v1.0-compliant theme file in tests
This took me quite a while because I couldn't make custom tags work with
the yaml package. I was referencing its integration on the
yaml-language-service package but I couldn't make heads or tails of it.

It carries quite a lot of extra baggage they internally use, that they
need because they have to make custom tags generically, for all possible
user configurations.