Compile schemas for the validation test
This took quite a while.

There's a lot of things in the VSCode schemas for color themes that
doesn't quite comply with the JSON Schema spec, or AJV's approach at
least.

I should extract schema compilation and the theme file parsing into
setup hooks or fixtures, but that's for a later step.

I also fear that passing AJV validation may not be sufficient to prove
that the schema works correctly under yaml-language-server (our actual
target) and its dependency vscode-json-languageservice.

Maybe I'll try integrating that instead of AJV later on. The
eemeli/yaml integration makes sense though, if we increase the scope of
this project to provice a package to transform YAML themes into the
actual delivered JSON version.