Provide stricter validation for !alpha custom tags
This was the feature that made me decide to extract this whole patched
JSON Schema thing into its own project, when I was banging my head
against the wall trying to change it without tests and anything. [^1]

With this, validation for YAML vscode themes like Dracula or Lucario is
as strict as it can get, the only part not really validated is the
custom tag itself, which is stripped before it gets to the JSON Schema
validation.

I should definitely refactor tests to make them tighter, there's a lot
of repeated, boilerplate code. Probably add a custom matcher or two. But
for now, this Just Works.

[^1]: https://github.com/victor-gp/lucario-vscode-theme/commit/13c4b4c9fb9f22d69204d6ab1e12ed2d3e2736ba