Convert yaml-*-defs.yaml to .json for public usage
This JSON (sub) Schema written in YAML has worked so far, because whatever the
libraries used by VS Code and the yaml-language-server can process it.
But there's no need to try our luck like that when we can easily convert
to JSON.

I'll keep the YAML version as input, because it _is_ useful for writing
the schema, considering JSON is such a PITA.

This commit updates tests to the new reference, but does not patch
vscode schemas yet. I'll do that in the next commit because that's too
much noise. This means that tests are broken for this commit.